fix(Teleport): ensure mesh teleporting sets y position correctly
On some occasions, the mesh can have a world position higher than
the actual pointer is colliding with, therefore it's better to use
the local position of the target object as this seems to work more
accurately in all situations.

The terrain height check should also ensure negative height terrains
adhere to the correct destination height as well.